build kernel modules for all kernels
partial-jira-bug: CEM-17259

+++ containers/vrouter/kernel-build-init/entrypoint.sh
@@ -18,25 +18,42 @@ if [ ! -d "/usr/src/linux-headers-$kver" ] ; then
   exit 1
 fi
 
-vrouter_dir="/usr/src/vrouter-${contrail_version}"
-mkdir -p $vrouter_dir
-cp -ap /vrouter_src/. ${vrouter_dir}/
-chmod -R 755  ${vrouter_dir}
-rm -rf /vrouter_src
-templ=$(cat /opt/contrail/src/dkms.conf)
-content=$(eval "echo \"$templ\"")
-echo "$content" > $vrouter_dir/dkms.conf
+if [ ! -f lib/modules/$kver/updates/dkms/vrouter.ko ] ; then
+    vrouter_dir="/usr/src/vrouter-${contrail_version}"
+    mkdir -p $vrouter_dir
+    cp -ap /vrouter_src/. ${vrouter_dir}/
+    chmod -R 755  ${vrouter_dir}
+    rm -rf /vrouter_src
+    templ=$(cat /opt/contrail/src/dkms.conf)
+    content=$(eval "echo \"$templ\"")
+    echo "$content" > $vrouter_dir/dkms.conf
 
-mkdir -p /vrouter/${contrail_version}/build/include/
-mkdir -p /vrouter/${contrail_version}/build/dp-core
-dkms --verbose add -m vrouter -v "${contrail_version}"
-dkms --verbose build -m vrouter -v "${contrail_version}"
-cat /var/lib/dkms/vrouter/${contrail_version}/build/make.log
-dkms --verbose install -m vrouter -v "${contrail_version}"
-depmod -a
+    mkdir -p /vrouter/${contrail_version}/build/include/
+    mkdir -p /vrouter/${contrail_version}/build/dp-core
+    dkms --verbose add -m vrouter -v "${contrail_version}"
+    dkms --verbose build -m vrouter -v "${contrail_version}"
+    cat /var/lib/dkms/vrouter/${contrail_version}/build/make.log
+    dkms --verbose install -m vrouter -v "${contrail_version}"
+    depmod -a
+fi
 
 # check vrouter.ko was built
 ls -l /lib/modules/$kver/updates/dkms/vrouter.ko || exit 1
+
+kernel_modules=$(ls /lib/modules)
+for kver in $kernel_modules ; do
+    # vrouter doesn't support kernels version 5, remove check after fix
+    if ! [[ "$kver" =~ ^5. ]] ; then
+        if [ ! -f "/lib/modules/$kver/updates/dkms/vrouter.ko" ]; then
+            echo "DKMS auto installing for kernel $kver"
+            dkms autoinstall -k $kver
+            depmod -a
+
+            # check vrouter.ko was built
+            ls -l /lib/modules/$kver/updates/dkms/vrouter.ko || exit 1
+        fi
+    fi
+done
 
 touch $vrouter_dir/module_compiled
